FW-269. set usage to 100% if no tx cells in schedule.

+++ openstack/02b-MAChigh/pid.c
@@ -43,7 +43,8 @@ int16_t pid_compute_usageOfCell() {
     uint8_t i;
     int16_t returnVal;
     // what we desired is no packet in the queue buffer (0) 
-    pid_vars.prevError = schedule_getCellUsage()-TARGET_CELL_USAGE;
+    pid_vars.prevError = (int16_t)schedule_getCellUsage()-(int16_t)TARGET_CELL_USAGE;
+    
     // udpate the history, elment zero is the newest one
     for (i=0;i<PID_NUM_INTEGRAL_HISTORY-1;i++) {
         pid_vars.errorHistory[PID_NUM_INTEGRAL_HISTORY-i-1] = pid_vars.errorHistory[PID_NUM_INTEGRAL_HISTORY-i-2];
